+++ NEWS.md
@@ -1,6 +1,17 @@
 # `unitizer` NEWS
 
-## v1.4.13
+## v1.4.15
+
+* Internal tests switched to `aammrtf`.
+* Upgrade process modified so that old `unitizer` that pass tests are not
+  required to be upgraded.
+* `show.progress` parameter added to control chattiness of progress updates.
+* Invalid `unitizers` now prevent review of valid `unitizers`.  Fix the invalid
+  `unitizers`, or remove them from the review list (e.g. by moving them out of
+  the primary `unitizer` directory).
+* `review` exposes `use.diff` and `show.progress` parameters.
+
+## v1.4.13-14
 
 * `unitizer` no longer directly accesses `.Traceback` to set traces in
   interactive sessions.  Instead, `traceback()` and `.traceback()` are masked